+++ iksTerm/README.md
@@ -60,9 +60,9 @@ In basic version we have:
 
 ## Future Development
 
-- Process control sequences & signals
-- UTF-8 except of ASCII
-- Keep history and get it by arrows
-- More keys to move cursor(`\b`)
-- Resizing
-- Custom font upload(improve)
+- [x] Resizing
+- [ ] Custom font upload(improve)
+- [ ] Process control sequences & signals
+- [ ] UTF-8 except of ASCII
+- [ ] Keep history and get it by arrows
+- [ ] More keys to move cursor(`\b`)
